chore: move MenuPanel into its own component
